+++ src/azure-cli/HISTORY.rst
@@ -15,6 +15,7 @@ Release History
 **AppConfig**
 
 * Add support for importing/exporting feature flags
+* Add new command 'az appconfig kv set-keyvault' for creating keyvault reference
 
 **AppService**